Config and Details Class Separated
