+++ mock-mpi/src/test/resources/vista-sites.json
@@ -1,0 +1,7 @@
+{
+  "default" : {
+    "sites" : [
+      "123"
+    ]
+  }
+}
